Fixed end blend chain joint translation and rotation
in blendChainConstraints() - account for the appropriate joint (ankle/wrist) and skip translations along all axis

blendTranslations() - take the values of Ik joint and static length of the joint, and use blendTwoAttr node to blend between the two and drive .translateX of the blend ankle/wrist joint.

Arm build commented out for debugging.