Foot Bank/ Orient to Parent
Finally implemented changes as discussed with V in relation to introducing banking in the foot system, as well as orient constraints for the head and limbs to provide control over whether these are oriented to world or to their parent.

Minor clean-up of unnecessary additional knots in the shapes file as all continuous shapes atm are created from a circle with 8 cvs and therefore all knots information is identical.